Meetings bugfix; Refactor; JPQL for meeting-participants requests

+++ src/main/kotlin/com/planify/planifyspring/main/features/meetings/routing/dto/get_my_meetings_short/GetMyMeetingsShortResponseDTO.kt
@@ -3,5 +3,5 @@ package com.planify.planifyspring.main.features.meetings.routing.dto.get_my_meet
 import java.time.Instant
 
 data class GetMyMeetingsShortResponseDTO(
-    val meetings: Map<Instant, Int>
+    val meetings: Map<Instant, Long>
 )
